WebVTT percentage value can be a float https://bugs.webkit.org/show_bug.cgi?id=134511
Reviewed by Brent Fulgham.


Source/WebCore:
No new tests, existing tests modified.

* html/track/TextTrackCueGeneric.cpp:
(WebCore::TextTrackCueGeneric::setLine): int -> double.
(WebCore::TextTrackCueGeneric::setPosition): Ditto.
* html/track/TextTrackCueGeneric.h:

* html/track/VTTCue.cpp:
(WebCore::VTTCue::setLine): Ditto.
(WebCore::VTTCue::setPosition):  Ditto.
(WebCore::VTTCue::setCueSettings): Support floating point values for line, position, and size.
(WebCore::scanPercentage): Deleted.
* html/track/VTTCue.h:

* html/track/VTTScanner.cpp:
(WebCore::VTTScanner::scanFloat): Support negative values.
* html/track/VTTScanner.h:

* html/track/WebVTTParser.cpp: Include parseFloatPercentageValue in non-WEBVTT_REGIONS builds.
* html/track/WebVTTParser.h:

* platform/graphics/avfoundation/InbandTextTrackPrivateAVF.cpp:
(WebCore::InbandTextTrackPrivateAVF::processCueAttributes): Drive by change to take a cue
    reference rather than a pointer.
(WebCore::InbandTextTrackPrivateAVF::processAttributedStrings): Ditto.
(WebCore::InbandTextTrackPrivateAVF::removeCompletedCues): Loop over all cues, don't LOG that
    we are asking the client to delete the cue because the client will have already deleted
    it when status when to GenericCueData::Complete.
* platform/graphics/avfoundation/InbandTextTrackPrivateAVF.h:

LayoutTests:
* media/track/captions-webvtt/tc015-positioning-ltr.vtt:
* media/track/captions-webvtt/tc015-positioning.vtt:
* media/track/captions-webvtt/tc017-line-position.vtt:
* media/track/captions-webvtt/tc018-align-text-line-position.vtt:
* media/track/captions-webvtt/tc019-cue-size.vtt:
* media/track/track-webvtt-tc015-positioning-expected.txt:
* media/track/track-webvtt-tc015-positioning.html:
* media/track/track-webvtt-tc017-line-position-expected.txt:
* media/track/track-webvtt-tc017-line-position.html:
* media/track/track-webvtt-tc018-align-text-line-position-expected.txt:
* media/track/track-webvtt-tc018-align-text-line-position.html:
* media/video-test.js:
(testCues):


Canonical link: https://commits.webkit.org/152522@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@170694 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/html/track/TextTrackCueGeneric.cpp
@@ -121,13 +121,13 @@ PassRefPtr<VTTCueBox> TextTrackCueGeneric::createDisplayTree()
     return TextTrackCueGenericBoxElement::create(ownerDocument(), *this);
 }
 
-void TextTrackCueGeneric::setLine(int line, ExceptionCode& ec)
+void TextTrackCueGeneric::setLine(double line, ExceptionCode& ec)
 {
     m_defaultPosition = false;
     VTTCue::setLine(line, ec);
 }
 
-void TextTrackCueGeneric::setPosition(int position, ExceptionCode& ec)
+void TextTrackCueGeneric::setPosition(double position, ExceptionCode& ec)
 {
     m_defaultPosition = false;
     VTTCue::setPosition(position, ec);
